+++ my-projects/practice-4/src/thread_pool/naive.rs
@@ -11,7 +11,7 @@ use super::ThreadPool;
 pub struct NaiveThreadPool;
 
 impl ThreadPool for NaiveThreadPool {
-    fn new(_threads: u32) -> Result<Self> {
+    fn new(_threads: usize) -> Result<Self> {
         Ok(NaiveThreadPool)
     }
